#242| +Shibha | Minor changes shipment file processing.

+++ modules/shipment/src/main/java/org/openlmis/shipment/domain/ShippedLineItem.java
@@ -23,7 +23,7 @@ public class ShippedLineItem extends BaseModel implements Importable{
   @ImportField(mandatory = true, name = "Product Code")
   private String productCode;
 
-  @ImportField(mandatory = true, type = "int", name = "Quantity Shipped")
+  @ImportField(mandatory = true, type = "double", name = "Quantity Shipped")
   private Integer quantityShipped;
 
 }
